Close uTP connection if origin down
When the injector received an EOF from origin (if it was down) it did not
clost the uTP connection it had with the injector_helper. Thus injector_helper
kept waiting for a HTTP response until it timed out after 50 seconds. This
commit closes the uTP connection and upates test to make sure the client
is notified quickly about the origin being down.